# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Marlin Frozen Fruits ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /materials:
    get:
      operationId: listMaterials
      tags: [materials]
      summary: List all materials
      responses:
        "200":
          description: Materials list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Material"
    post:
      operationId: createMaterial
      tags: [materials]
      summary: Create material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
  /materials/{id}:
    get:
      operationId: getMaterial
      tags: [materials]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
    patch:
      operationId: updateMaterial
      tags: [materials]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MaterialUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Material"
    delete:
      operationId: deleteMaterial
      tags: [materials]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # ── RAW MATERIALS ─────────────────────────────────────────────────────────
  /raw-materials:
    get:
      operationId: listRawMaterials
      tags: [raw-materials]
      summary: List all raw materials
      responses:
        "200":
          description: Raw materials list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RawMaterial"
    post:
      operationId: createRawMaterial
      tags: [raw-materials]
      summary: Create raw material
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RawMaterialInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RawMaterial"
  /raw-materials/{id}:
    get:
      operationId: getRawMaterial
      tags: [raw-materials]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Raw material
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RawMaterial"
    patch:
      operationId: updateRawMaterial
      tags: [raw-materials]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RawMaterialUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RawMaterial"
    delete:
      operationId: deleteRawMaterial
      tags: [raw-materials]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # ── ITEM MASTER ───────────────────────────────────────────────────────────
  /items:
    get:
      operationId: listItems
      tags: [item-master]
      summary: List all items in item master
      responses:
        "200":
          description: Items list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Item"
    post:
      operationId: createItem
      tags: [item-master]
      summary: Create item in master
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
  /items/{id}:
    get:
      operationId: getItem
      tags: [item-master]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    patch:
      operationId: updateItem
      tags: [item-master]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Item"
    delete:
      operationId: deleteItem
      tags: [item-master]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "204":
          description: Deleted

  # ── PURCHASE ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Material:
      type: object
      required: [id, name, unit, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        unit:
          type: string
        description:
          type: ["string", "null"]
        currentStock:
          type: number
        createdAt:
          type: string
    MaterialInput:
      type: object
      required: [name, unit]
      properties:
        name:
          type: string
        unit:
          type: string
        description:
          type: string
    MaterialUpdate:
      type: object
      properties:
        name:
          type: string
        unit:
          type: string
        description:
          type: string

    # RAW MATERIAL
    RawMaterial:
      type: object
      required: [id, name, unit, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        unit:
          type: string
        description:
          type: ["string", "null"]
        currentStock:
          type: number
        createdAt:
          type: string
    RawMaterialInput:
      type: object
      required: [name, unit]
      properties:
        name:
          type: string
        unit:
          type: string
        description:
          type: string
    RawMaterialUpdate:
      type: object
      properties:
        name:
          type: string
        unit:
          type: string
        description:
          type: string

    # ITEM MASTER
    Item:
      type: object
      required: [id, name, hsnCode, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        hsnCode:
          type: string
        taxRate:
          type: number
        unit:
          type: string
        description:
          type: ["string", "null"]
        productionStock:
          type: number
        createdAt:
          type: string
    ItemInput:
      type: object
      required: [name, hsnCode, unit]
      properties:
        name:
          type: string
        hsnCode:
          type: string
        taxRate:
          type: number
        unit:
          type: string
        description:
          type: string
    ItemUpdate:
      type: object
      properties:
        name:
          type: string
        hsnCode:
          type: string
        taxRate:
          type: number
        unit:
          type: string
        description:
          type: string

    # PURCHASE
...